+++ client/src/lib/index.ts
@@ -3,3 +3,4 @@ export * from './conversation/index.js';
 export * from './room/index.js';
 export * from './API.js';
 export * from './types.js';
+export * from 'missiv';