+++ client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # missiv-client
 
+## 0.0.5
+
+### Patch Changes
+
+- MissivUser + types export for client
+
 ## 0.0.4
 
 ### Patch Changes