+++ server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # missiv-server
 
+## 0.0.4
+
+### Patch Changes
+
+- MissivUser + types export for client
+- Updated dependencies
+  - missiv@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes